Add pom.xml to build with Maven
If I'm completely honest I never figured out how to build this project
manually. I stopped learning Java years ago and always relied on the
Eclipse IDE to get the work done for me. I did however always feel bad
for anyone out there trying to build this project without having to
install a specific IDE.

Maven finally unlocked this for me. It took me about an hour of
tutorials and examples, but now building the project is easy as pie.

I also moved the files to adhere to the expected / recommended folder
structure for Maven.

I'm very happy that I finally got this right. I seem to come back to
this project every two years or so (the current trigger was a toy my
little kid got that plays an earily familiar cat sound) and trying to
build it up until now was always a pain in the ass.