Add default options.cfg
This file can be used to change some options without recompiling the
game. Most of the options are self-explainatory, but DynamicLightning
might be a bit confusing. There are 3 settings:

- 0 (off)
- 1 (create glow around some projectiles and explosions)
- 2 (add a glow around the protagonist as if he's using a flashlight)

The last setting is incredibly resource hungry and laggy. I'm pretty
sure that there is something wrong with the algorithm I used way back
when. Everything should work smoothly though at '1'.