Add Makefile
The Makefile makes it even easier to build and run the game. The
necessary command (in case you don't have make installed) can still be
found in the Makefile.

As mentioned in a previous commit, this is mostly for my benefit as I
return to this project every few years to play 'Miez Attack' again. The
easier it is to build, the more likely I am to enjoy my short time with
it.